+++ src/Smartstore.Core/Checkout/Rules/Impl/CustomerAuthenticationRule.cs
@@ -1,10 +1,20 @@
-﻿using Smartstore.Core.Data;
+﻿using System.Collections.Frozen;
+using Smartstore.Core.Data;
 using Smartstore.Core.Rules;
 
 namespace Smartstore.Core.Checkout.Rules.Impl
 {
     internal class CustomerAuthenticationRule : IRule<CartRuleContext>
     {
+        // INFO: ExternalAuthenticationRecord.ProviderSystemName always contains the old system name, even for new logins!
+        private readonly static FrozenDictionary<string, string> _legacyAuthenticationNamesMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase)
+        {
+            { "Smartstore.FacebookAuth", "Smartstore.Facebook.Auth" },
+            { "Smartstore.TwitterAuth", "Smartstore.Twitter.Auth" },
+            { "Smartstore.GoogleAuth", "Smartstore.Google.Auth" }
+        }
+        .ToFrozenDictionary(StringComparer.OrdinalIgnoreCase);
+
         private readonly SmartDbContext _db;
 
         public CustomerAuthenticationRule(SmartDbContext db)
@@ -21,7 +31,7 @@ namespace Smartstore.Core.Checkout.Rules.Impl
 
                 // INFO: checking whether the authentication methods are also active is probably not necessary here.
                 var authenticationMethods = customer.ExternalAuthenticationRecords
-                    .Select(x => MapLegacyAuthenticationName(x.ProviderSystemName))
+                    .Select(x => _legacyAuthenticationNamesMap.Get(x.ProviderSystemName.EmptyNull()) ?? x.ProviderSystemName)
                     .ToArray();
 
                 var match = expression.HasListsMatch(authenticationMethods, StringComparer.InvariantCultureIgnoreCase);
@@ -29,21 +39,6 @@ namespace Smartstore.Core.Checkout.Rules.Impl
             }
 
             return false;
-        }
-
-        private static string MapLegacyAuthenticationName(string systemName)
-        {
-            switch (systemName.NullEmpty().ToLowerInvariant())
-            {
-                case "smartstore.facebookauth":
-                    return "Smartstore.Facebook.Auth";
-                case "smartstore.twitterauth":
-                    return "Smartstore.Twitter.Auth";
-                case "smartstore.googleauth":
-                    return "Smartstore.Google.Auth";
-                default:
-                    return systemName;
-            }
         }
     }
 }